Langgraph框架/requirements.txt
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
python-dotenv==1.0.1
pydantic==2.9.2
typing-extensions==4.12.2
requests==2.32.3
beautifulsoup4==4.12.3
matplotlib==3.9.2
pandas==2.2.3
numpy==1.26.4
